+++ src/pages/index.ts
@@ -1,3 +1,2 @@
-export { default as Layout } from "./Layout"
 export { default as Nothing } from "./Nothing"
 export { default as Overview } from "./Overview"
